A small enhancement/adjustment.
In the tests and the examples in the docs in the system roles, there
is a roles dir containing a symbolic link to the top level of each role,
e.g., kernel_settings/tests/roles/linux-system-roles.kernel_settings -> ../../
Once converted to collection format, the symlink is not needed but there
is few exceptions, e.g., in this case referring the path in include_vars.
  include_vars:
    file: roles/linux-system-roles.kernel_settings/vars/main.yml
This patch keeps the symlink and path in the test playbook only when the
path is needed. A function symlink_n_rolename and its sub-functions take
care of it.

As a result, most symlinks are not included in the converted tree.
2 exceptions.
  - tests/kernel_settings/roles/linux-system-roles.kernel_settings ->
    ../../../roles/kernel_settings/
  - vars/CentOS_?.yml --> vars/RedHat_?.yml